fix: move event bridge and auto-start from Dashboard to AppLayout
